+++ src-java/base-topology/base-storm-topology/src/test/java/org/openkilda/wfm/config/ZookeeperConfig.java
@@ -23,8 +23,8 @@ import com.sabre.oss.conf4j.annotation.Key;
 @Key("zookeeper")
 public interface ZookeeperConfig {
     // To be able to run tests in parallel we must use different ports for zookeeper hosts in each test
-    int STATS_TOPOLOGY_TEST_ZOOKEEPER_PORT = 9093;
-    int ISL_LATENCY_TOPOLOGY_TEST_ZOOKEEPER_PORT = 9094;
+    int STATS_TOPOLOGY_TEST_ZOOKEEPER_PORT = 9098;
+    int ISL_LATENCY_TOPOLOGY_TEST_ZOOKEEPER_PORT = 9099;
 
     @Key("hosts")
     String getHosts();