+++ src-java/nbworker-topology/nbworker-storm-topology/src/main/java/org/openkilda/wfm/share/mappers/RequestedFlowMapper.java
@@ -15,20 +15,21 @@
 
 package org.openkilda.wfm.share.mappers;
 
+import org.openkilda.adapter.FlowDestAdapter;
+import org.openkilda.adapter.FlowSourceAdapter;
 import org.openkilda.messaging.command.flow.FlowRequest;
 import org.openkilda.model.Flow;
-import org.openkilda.model.FlowEndpoint;
 import org.openkilda.model.PathComputationStrategy;
 import org.openkilda.model.Switch;
 import org.openkilda.model.SwitchId;
 
+import lombok.NonNull;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 
-import java.util.Optional;
-
-@Mapper(imports = {FlowEndpoint.class, Switch.class, Optional.class, PathComputationStrategy.class})
+// TODO(surabujin): merge with org.openkilda.wfm.topology.flowhs.mapper.RequestedFlowMapper
+@Mapper
 public abstract class RequestedFlowMapper {
 
     public static final RequestedFlowMapper INSTANCE = Mappers.getMapper(RequestedFlowMapper.class);
@@ -36,16 +37,36 @@ public abstract class RequestedFlowMapper {
     /**
      * Convert {@link Flow} to {@link FlowRequest}.
      */
+    public FlowRequest toFlowRequest(Flow flow) {
+        FlowRequest request = generatedMap(flow);
+        request.setSource(new FlowSourceAdapter(flow).getEndpoint());
+        request.setDestination(new FlowDestAdapter(flow).getEndpoint());
+        return request;
+    }
+
+    /**
+     * Convert {@link FlowRequest} to {@link Flow}.
+     */
+    @Mapping(target = "srcSwitch", source = "source.switchId")
+    @Mapping(target = "srcPort", source = "source.portNumber")
+    @Mapping(target = "srcVlan", source = "source.outerVlanId")
+    @Mapping(target = "srcInnerVlan", source = "source.innerVlanId")
+    @Mapping(target = "destSwitch", source = "destination.switchId")
+    @Mapping(target = "destPort", source = "destination.portNumber")
+    @Mapping(target = "destVlan", source = "destination.outerVlanId")
+    @Mapping(target = "destInnerVlan", source = "destination.innerVlanId")
+    @Mapping(target = "detectConnectedDevices", source = "detectConnectedDevices")
+    @Mapping(target = "diverseGroupId", ignore = true)
+    @Mapping(target = "affinityGroupId", ignore = true)
+    @Mapping(target = "statusInfo", ignore = true)
+    @Mapping(target = "targetPathComputationStrategy", ignore = true)
+    @Mapping(target = "status", ignore = true)
+    @Mapping(target = "yFlowId", ignore = true)
+    @Mapping(target = "yFlow", ignore = true)
+    public abstract Flow toFlow(@NonNull FlowRequest request);
+
     @Mapping(target = "flowId", source = "flowId")
-    @Mapping(target = "source", expression = "java(new FlowEndpoint(flow.getSrcSwitchId(), "
-            + "flow.getSrcPort(), flow.getSrcVlan(), flow.getSrcInnerVlan()))")
-    @Mapping(target = "destination", expression = "java(new FlowEndpoint(flow.getDestSwitchId(), "
-            + "flow.getDestPort(), flow.getDestVlan(), flow.getDestInnerVlan()))")
     @Mapping(target = "encapsulationType", source = "encapsulationType")
-    @Mapping(target = "pathComputationStrategy",
-            expression = "java(java.util.Optional.ofNullable(flow.getPathComputationStrategy())"
-                    + ".map(pcs -> pcs.toString().toLowerCase())"
-                    + ".orElse(null))")
     @Mapping(target = "bandwidth", source = "bandwidth")
     @Mapping(target = "ignoreBandwidth", source = "ignoreBandwidth")
     @Mapping(target = "periodicPings", source = "periodicPings")
@@ -55,47 +76,36 @@ public abstract class RequestedFlowMapper {
     @Mapping(target = "priority", source = "priority")
     @Mapping(target = "pinned", source = "pinned")
     @Mapping(target = "detectConnectedDevices", source = "detectConnectedDevices")
+    @Mapping(target = "loopSwitchId", source = "loopSwitchId")
+    @Mapping(target = "source", ignore = true)
+    @Mapping(target = "destination", ignore = true)
     @Mapping(target = "transitEncapsulationId", ignore = true)
     @Mapping(target = "diverseFlowId", ignore = true)
     @Mapping(target = "affinityFlowId", ignore = true)
     @Mapping(target = "type", ignore = true)
     @Mapping(target = "bulkUpdateFlowIds", ignore = true)
     @Mapping(target = "doNotRevert", ignore = true)
-    public abstract FlowRequest toFlowRequest(Flow flow);
-
-    /**
-     * Convert {@link FlowRequest} to {@link Flow}.
-     */
-    @Mapping(target = "srcSwitch", expression = "java(java.util.Optional.ofNullable(request.getSource())"
-            + ".map(FlowEndpoint::getSwitchId).map(id -> Switch.builder().switchId(id).build()).orElse(null))")
-    @Mapping(target = "srcPort", expression = "java(java.util.Optional.ofNullable(request.getSource())"
-            + ".map(FlowEndpoint::getPortNumber).orElse(null))")
-    @Mapping(target = "srcVlan", expression = "java(java.util.Optional.ofNullable(request.getSource())"
-            + ".map(FlowEndpoint::getOuterVlanId).orElse(null))")
-    @Mapping(target = "srcInnerVlan", expression = "java(java.util.Optional.ofNullable(request.getSource())"
-            + ".map(FlowEndpoint::getInnerVlanId).orElse(null))")
-    @Mapping(target = "destSwitch", expression = "java(java.util.Optional.ofNullable(request.getDestination())"
-            + ".map(FlowEndpoint::getSwitchId).map(id -> Switch.builder().switchId(id).build()).orElse(null))")
-    @Mapping(target = "destPort", expression = "java(java.util.Optional.ofNullable(request.getDestination())"
-            + ".map(FlowEndpoint::getPortNumber).orElse(null))")
-    @Mapping(target = "destVlan", expression = "java(java.util.Optional.ofNullable(request.getDestination())"
-            + ".map(FlowEndpoint::getOuterVlanId).orElse(null))")
-    @Mapping(target = "destInnerVlan", expression = "java(java.util.Optional.ofNullable(request.getDestination())"
-            + ".map(FlowEndpoint::getInnerVlanId).orElse(null))")
-    @Mapping(target = "pathComputationStrategy",
-            expression = "java(Optional.ofNullable(request.getPathComputationStrategy())"
-                    + ".map(String::toUpperCase).map(PathComputationStrategy::valueOf).orElse(null))")
-    @Mapping(target = "detectConnectedDevices", source = "detectConnectedDevices")
-    @Mapping(target = "diverseGroupId", ignore = true)
-    @Mapping(target = "affinityGroupId", ignore = true)
-    @Mapping(target = "statusInfo", ignore = true)
-    @Mapping(target = "targetPathComputationStrategy", ignore = true)
-    @Mapping(target = "status", ignore = true)
-    @Mapping(target = "yFlowId", ignore = true)
-    @Mapping(target = "yFlow", ignore = true)
-    public abstract Flow toFlow(FlowRequest request);
+    protected abstract FlowRequest generatedMap(Flow flow);
 
     public SwitchId map(String value) {
         return value == null ? null : new SwitchId(value);
+    }
+
+    protected String mapPathComputationStrategy(PathComputationStrategy strategy) {
+        if (strategy == null) {
+            return null;
+        }
+        return strategy.toString().toLowerCase();
+    }
+
+    protected PathComputationStrategy mapPathComputationStrategy(String raw) {
+        if (raw == null) {
+            return null;
+        }
+        return PathComputationStrategy.valueOf(raw.toUpperCase());
+    }
+
+    protected Switch newSwitch(SwitchId switchId) {
+        return Switch.builder().switchId(switchId).build();
     }
 }