+++ src-java/northbound-service/northbound-api/src/main/java/org/openkilda/northbound/dto/v2/yflows/YFlowPingPayload.java
@@ -1,4 +1,4 @@
-/* Copyright 2018 Telstra Open Source
+/* Copyright 2022 Telstra Open Source
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -13,10 +13,21 @@
  *   limitations under the License.
  */
 
-package org.openkilda.wfm.topology.ping.bolt;
+package org.openkilda.northbound.dto.v2.yflows;
 
-import org.openkilda.wfm.share.bolt.KafkaDecoder;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
 
-public class InputDecoder extends KafkaDecoder {
-    public static final String BOLT_ID = ComponentId.INPUT_DECODER.toString();
+import javax.validation.constraints.PositiveOrZero;
+
+@Data
+@Builder
+@AllArgsConstructor
+public class YFlowPingPayload {
+
+    @PositiveOrZero(message = "timeoutMillis can't be negative")
+    @JsonProperty("timeout")
+    int timeoutMillis;
 }