+++ src-java/northbound-service/northbound/src/test/java/org/openkilda/northbound/service/SwitchServiceTest.java
@@ -29,24 +29,6 @@ import org.openkilda.messaging.info.switches.SwitchSyncResponse;
 import org.openkilda.model.SwitchId;
 import org.openkilda.northbound.MessageExchanger;
 import org.openkilda.northbound.config.KafkaConfig;
-import org.openkilda.northbound.converter.ConnectedDeviceMapper;
-import org.openkilda.northbound.converter.ConnectedDeviceMapperImpl;
-import org.openkilda.northbound.converter.FlowEncapsulationTypeMapper;
-import org.openkilda.northbound.converter.FlowEncapsulationTypeMapperImpl;
-import org.openkilda.northbound.converter.FlowMapper;
-import org.openkilda.northbound.converter.FlowMapperImpl;
-import org.openkilda.northbound.converter.FlowStatusMapper;
-import org.openkilda.northbound.converter.FlowStatusMapperImpl;
-import org.openkilda.northbound.converter.InstantMapper;
-import org.openkilda.northbound.converter.InstantMapperImpl;
-import org.openkilda.northbound.converter.LagPortMapper;
-import org.openkilda.northbound.converter.LagPortMapperImpl;
-import org.openkilda.northbound.converter.PathComputationStrategyMapper;
-import org.openkilda.northbound.converter.PathComputationStrategyMapperImpl;
-import org.openkilda.northbound.converter.PortPropertiesMapper;
-import org.openkilda.northbound.converter.PortPropertiesMapperImpl;
-import org.openkilda.northbound.converter.SwitchMapper;
-import org.openkilda.northbound.converter.SwitchMapperImpl;
 import org.openkilda.northbound.dto.v1.switches.RulesSyncDto;
 import org.openkilda.northbound.dto.v1.switches.RulesSyncResult;
 import org.openkilda.northbound.dto.v1.switches.SwitchSyncResult;
@@ -60,6 +42,7 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -147,6 +130,7 @@ public class SwitchServiceTest {
 
     @TestConfiguration
     @Import(KafkaConfig.class)
+    @ComponentScan({"org.openkilda.northbound.converter"})
     @PropertySource({"classpath:northbound.properties"})
     static class Config {
         @Bean
@@ -158,51 +142,5 @@ public class SwitchServiceTest {
         public SwitchService switchService(MessagingChannel messagingChannel) {
             return new SwitchServiceImpl(messagingChannel);
         }
-
-        @Bean
-        public SwitchMapper switchMapper() {
-            return new SwitchMapperImpl();
-        }
-
-        @Bean
-        public ConnectedDeviceMapper connectedDeviceMapper() {
-            return new ConnectedDeviceMapperImpl();
-        }
-
-        @Bean
-        public PortPropertiesMapper portPropertiesMapper() {
-            return new PortPropertiesMapperImpl();
-        }
-
-        @Bean
-        public InstantMapper instantMapper() {
-            return new InstantMapperImpl();
-        }
-
-        @Bean
-        public PathComputationStrategyMapper pathComputationStrategyMapper() {
-            return new PathComputationStrategyMapperImpl();
-        }
-
-        @Bean
-        public FlowStatusMapper flowStatusMapper() {
-            return new FlowStatusMapperImpl();
-        }
-
-        @Bean
-        public FlowEncapsulationTypeMapper flowEncapsulationTypeMapper() {
-            return new FlowEncapsulationTypeMapperImpl();
-        }
-
-        @Bean
-        public FlowMapper flowMapper() {
-            return new FlowMapperImpl();
-        }
-
-        @Bean
-        public LagPortMapper lagMapper() {
-            return new LagPortMapperImpl();
-        }
     }
-
 }